+++ src/main/java/net/silentchaos512/gear/crafting/recipe/IGearRecipe.java
@@ -4,28 +4,37 @@ import net.minecraft.world.Container;
 import net.minecraft.world.item.ItemStack;
 import net.silentchaos512.gear.api.item.GearType;
 import net.silentchaos512.gear.api.item.ICoreItem;
+import net.silentchaos512.gear.api.material.IMaterial;
 import net.silentchaos512.gear.api.material.IMaterialInstance;
 import net.silentchaos512.gear.api.part.PartType;
+import net.silentchaos512.gear.config.Config;
 import net.silentchaos512.gear.gear.material.MaterialInstance;
 import net.silentchaos512.gear.gear.part.PartData;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 interface IGearRecipe {
     ICoreItem getOutputItem();
 
     default Collection<PartData> getParts(Container inv) {
         List<MaterialInstance> materials = new ArrayList<>();
+        IMaterial first = null;
         List<PartData> parts = new ArrayList<>();
 
         for (int i = 0; i < inv.getContainerSize(); ++i) {
             ItemStack stack = inv.getItem(i);
+
             if (!stack.isEmpty()) {
                 MaterialInstance mat = MaterialInstance.from(stack);
+
                 if (mat != null) {
+                    // If classic mixing is disabled, all materials must be the same
+                    if (first == null) {
+                        first = mat.get();
+                    } else if (!Config.Common.allowLegacyMaterialMixing.get() && first != mat.get()) {
+                        return Collections.emptyList();
+                    }
+
                     materials.add(mat);
                 } else {
                     PartData part = PartData.from(stack);